labs/03-langchain/demo-04-max-token/pyproject.toml
[project]
name = "demo-4-max-token"
version = "0.1.0"
description = "Max Tokens Limit Demo: LangChain + Gemini with Token Limiting"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.0.0",
    "langchain-openai==1.0.0",
    "langchain-core==1.0.1",
    "python-dotenv==1.2.1",
    "fastapi==0.121.0",
    "uvicorn==0.38.0",
    "pydantic==2.12.4",
]
